Added comment describing `for...else` block
I saw this odd-looking code block structure, and was curious.  So I looked it up, and voila!  I discovered that it’s useful for eliminating flag variables in loops.

I’m pretty sure that using `for...else` is uncommon, however.  So, I linked to the blog article that enlightened me, to help anyone else who might read the code and scratch their head. :)